Fixing a test failure on Win7
This test specifically tests triggers with regular expressions where it looks for changes in root directrory in one trigger and another trigger monitoring any change in any sub-directory (recursive).

On Win7 when a file is modified we see a directory changed event is triggered on the parent directory automatically. This results in a flaky run on Win7. So making the test bit more restrictive to fix flakiness on Win7.